Make letterbox its own individual plugin
