+++ src/ui/letterbox/constants.rs
@@ -1,0 +1,1 @@
+pub const LETTERBOX_HEIGHT: f32 = 100.0;
