+++ src/ui/letterbox/resources.rs
@@ -1,0 +1,18 @@
+use bevy::prelude::{Component, Resource};
+
+#[derive(Component)]
+pub struct LetterboxTop;
+
+#[derive(Component)]
+pub struct LetterboxBottom;
+
+#[derive(Component)]
+pub struct LetterboxSide;
+
+#[derive(Component)]
+pub struct Letterbox;
+
+#[derive(Resource, Default)]
+pub struct LetterboxState {
+    pub active: bool,
+}
